fix: GcpFirestoreEmulatorAutoConfiguration constructs invalid document parent (#2348)
Fixes: #2290.
Fixes: #2330.
